refactor: rename to OnyxNotificationDot
